requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
pytest
pytest-asyncio
httpx
redis
celery
prometheus-client
python-json-logger


